Fixed Clang warnings about int conversions

+++ reverse_string.c
@@ -15,7 +15,7 @@ char *reverse_string(char *str) {
     }
     */
     int len = (int) strlen(str);
-    char *tstr = (char *) malloc(sizeof(char) * (int) strlen(str));
+    char *tstr = (char *) malloc(sizeof(char) * strlen(str));
     int i; /* Loop iterator */
     if (tstr == NULL)
         exit(EXIT_FAILURE);
